requirements.txt
Flask
google-generativeai
requests
gunicorn